requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
dotenv==0.9.9
faiss-cpu==1.13.2
fastapi==0.131.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
mysql-connector-python==9.6.0
numpy==2.4.2
openai==2.21.0
packaging==26.0
pydantic==2.12.5
pydantic_core==2.41.5
pyodbc==5.3.0
python-dotenv==1.2.1
sniffio==1.3.1
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
